refactor hap clust
